# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GYMCO customer API — multi-gym memberships, classes, bookings, check-ins, and fitness tracking.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/reviews/{id}/moderation:
    post:
      operationId: moderateTrainerReview
      tags: [reviews]
      security: [{adminSession: []}]
      parameters:
        - {in: path, name: id, required: true, schema: {type: integer, minimum: 1}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/TrainerReviewModerationInput"}
      responses:
        "200":
          description: Moderated member review
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Review"}
...
  /trainers:
    get:
      operationId: listTrainers
      tags: [trainers]
      summary: Browse trainers in the marketplace
      parameters:
        - in: query
          name: specialty
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Trainer" }
  /trainers/{trainerId}:
    get:
      operationId: getTrainer
      tags: [trainers]
      summary: Trainer profile
      parameters:
        - in: path
          name: trainerId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Trainer" }
...
components:
  securitySchemes:
    memberBearer:
      type: http
      scheme: bearer
    adminSession:
      type: apiKey
      in: cookie
      name: connect.sid
...
  schemas:
...
    TrainerReviewModerationInput:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [approved, rejected]}
...
    Review:
      type: object
      required: [id, reviewerName, branchName, reviewText, rating, isSample, isPublished, sortOrder, createdAt, updatedAt]
      properties:
        isMemberReview: {type: boolean}
        moderationStatus: {type: [string, "null"], enum: [pending, approved, rejected, null]}
        id: {type: integer}
        reviewerName: {type: string}
        branchName: {type: string}
        trainerId: {type: [string, "null"]}
        gymId: {type: [integer, "null"]}
        reviewText: {type: string}
        rating: {type: integer, minimum: 1, maximum: 5}
        isSample: {type: boolean}
        isPublished: {type: boolean}
        sortOrder: {type: integer}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
...
    Trainer:
      type: object
      required:
        [
          id, name, specialty, bio, photoUrl, rating, sessionsCount,
          pricePerSession, certifications, city
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        specialty: { type: string }
        bio: { type: string }
        photoUrl: { type: string }
        rating: { type: number }
        sessionsCount: { type: integer }
        pricePerSession: { type: integer }
        certifications:
          type: array
          items: { type: string }
        city: { type: string }
...